refactor(blog): extract shared markdown renderer from Vite plugins
